Add endGame() and resetGame() function and condensed playRound()
Condensed down the playRound() so it does the same thing with less code.
Add updateScore() so that it updates and displays the score, while also checking if it should run the endGame() function.
Add endGame() function that displays final message and disables the buttons, so no further round can be played.
Add resetGame() so that score and text can be reset, and the buttons are re-enabled.